+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [3.X.X] - 2022-X-X
 ### Added
+- rich schema script for specific datasets. [RW-141](https://vizzuality.atlassian.net/browse/RW-141)
 
 ### Changed
 - dataset fetching determined by `NEXT_PUBLIC_ENVS_SHOW` environmental variable.
@@ -15,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - bump `next@12.1.6`.
 
 ### Fixed
+- map overlay style that prevented a user could draw on the map.
+- typo in published layers checkbox. [RW-146](https://vizzuality.atlassian.net/browse/RW-146)
 - Explore: fixes navigation when a user clicks on "All datasets" button after searching. Now the search results remain. [RW-87](https://vizzuality.atlassian.net/browse/RW-87)
 
 ### Removed